modules/segmentation/requirements.txt
# Semantic segmentation (Pytorch-UNet) — install in a venv alongside root requirements if needed
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.21.0
Pillow>=9.0.0
tqdm>=4.64.0
PyYAML>=6.0
